requirements.txt
plaintext
instagrapi
python-dotenv
schedule
